+++ lua/cargdev/plugins/lazygit.lua
@@ -1,5 +1,4 @@
 return {
-  --[[ dir = "/Users/carlos/Documents/SSD_Documents/projects/lazygit.nvim", ]]
   "kdheepak/lazygit.nvim",
   cmd = {
     "LazyGit",